+++ bidscoin/cli/_bidsparticipants.py
@@ -22,9 +22,9 @@ def get_parser():
                                      epilog='examples:\n'
                                             '  bidsparticipants myproject/raw myproject/bids\n'
                                             '  bidsparticipants myproject/raw myproject/bids -b mybidsmap\n ')
-    parser.add_argument('sourcefolder',     help='The study root folder containing the raw source data folders')
+    parser.add_argument('sourcefolder',     help='The dataset root folder containing the raw source data folders')
     parser.add_argument('bidsfolder',       help='The destination/output folder with the bids data')
     parser.add_argument('-d','--dryrun',    help='Do not save anything, only print the participants info on screen', action='store_true')
-    parser.add_argument('-b','--bidsmap',   help='The study bidsmap file with the mapping heuristics. If the bidsmap filename is just the base name (i.e. no "/" in the name) then it is assumed to be located in the current directory or in bidsfolder/code/bidscoin. Default: bidsmap.yaml', metavar='NAME', default='bidsmap.yaml')
+    parser.add_argument('-b','--bidsmap',   help='The dataset bidsmap file with the mapping heuristics. If the bidsmap filename is just the base name (i.e. no "/" in the name) then it is assumed to be located in the current directory or in bidsfolder/code/bidscoin. Default: bidsmap.yaml', metavar='NAME', default='bidsmap.yaml')
 
     return parser